+++ libs/sdk-core/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.11.2 (unreleased)
+
+### Improvements
+
+- Add SAML related properties to `Account` interface
+- Add `AccountModification` interface
+
 # 1.11.1 (2024-01-22)
 
 ### Improvements